+++ apps/app/src/features/questionnaire/client/components/Question.tsx
@@ -1,5 +1,6 @@
-import { IQuestionHasId } from '~/interfaces/questionnaire/question';
 import { useCurrentUser } from '~/stores/context';
+
+import { IQuestionHasId } from '../../interfaces/question';
 
 type QuestionProps = {
   question: IQuestionHasId,
