+++ apps/app/src/features/questionnaire/client/stores/questionnaire.tsx
@@ -1,7 +1,8 @@
 import useSWR, { SWRResponse } from 'swr';
 
 import { apiv3Get } from '~/client/util/apiv3-client';
-import { IQuestionnaireOrderHasId } from '~/interfaces/questionnaire/questionnaire-order';
+
+import { IQuestionnaireOrderHasId } from '../../interfaces/questionnaire-order';
 
 export const useSWRxQuestionnaireOrders = (): SWRResponse<IQuestionnaireOrderHasId[], Error> => {
   return useSWR(
