+++ apps/app/src/features/questionnaire/interfaces/proactive-questionnaire-answer.ts
@@ -1,6 +1,7 @@
 import { IGrowiInfo } from './growi-info';
 import { IUserInfo } from './user-info';
 
+
 export interface IProactiveQuestionnaireAnswer {
   satisfaction: number,
   commentText: string,
