+++ apps/app/src/features/questionnaire/server/models/questionnaire-answer-status.ts
@@ -1,7 +1,8 @@
 import { Model, Schema, Document } from 'mongoose';
 
-import { IQuestionnaireAnswerStatus, StatusType } from '~/interfaces/questionnaire/questionnaire-answer-status';
 import { getOrCreateModel } from '~/server/util/mongoose-utils';
+
+import { IQuestionnaireAnswerStatus, StatusType } from '../../interfaces/questionnaire-answer-status';
 
 export interface QuestionnaireAnswerStatusDocument extends IQuestionnaireAnswerStatus, Document {}
