+++ apps/app/src/features/questionnaire/server/models/schema/answer.ts
@@ -1,7 +1,8 @@
 import { Schema } from 'mongoose';
 
-import { IAnswer } from '~/interfaces/questionnaire/answer';
 import { ObjectIdLike } from '~/server/interfaces/mongoose-utils';
+
+import { IAnswer } from '../../../interfaces/answer';
 
 export const answerSchema = new Schema<IAnswer<ObjectIdLike>>({
   question: { type: Schema.Types.ObjectId, ref: 'Question', required: true },
