+++ apps/app/src/server/routes/apiv3/index.js
@@ -112,7 +112,7 @@ module.exports = (crowi, app) => {
 
   router.use('/user-ui-settings', require('./user-ui-settings')(crowi));
 
-  router.use('/questionnaire', require('./questionnaire')(crowi));
+  router.use('/questionnaire', require('~/features/questionnaire/server/routes/apiv3/questionnaire')(crowi));
 
 
   return [router, routerForAdmin, routerForAuth];
